+++ glean-core/android/src/main/java/mozilla/telemetry/glean/rust/LibGleanFFI.kt
@@ -89,11 +89,17 @@ internal interface LibGleanFFI : Library {
 
     fun glean_initialize(cfg: FfiConfiguration): Byte
 
+    fun glean_flush_rlb_dispatcher()
+
     fun glean_clear_application_lifetime_metrics()
 
     fun glean_set_dirty_flag(flag: Byte)
 
     fun glean_is_dirty_flag_set(): Byte
+
+    fun glean_handle_client_active()
+
+    fun glean_handle_client_inactive()
 
     fun glean_test_clear_all_stores()