+++ glean-core/ios/GleanTests/Metrics/EventMetricTests.swift
@@ -38,7 +38,13 @@ class EventMetricTypeTests: XCTestCase {
     var lastPingJson: [String: Any]?
 
     private func setupHttpResponseStub() {
-        stubServerReceive { _, json in
+        stubServerReceive { pingType, json in
+            if pingType != "events" {
+                // Skip non-events pings here.
+                // This might include the initial "active" baseline ping.
+                return
+            }
+
             XCTAssert(json != nil)
             self.lastPingJson = json